+++ src/gas/routes.ts
@@ -1,4 +1,4 @@
-import Utils from "./shared/Utils";
+import Utils from "./shared/Response";
 import ParameterInterface from "./types/ParameterInterface";
 import ResponseInterface from "./types/ResponseInterface";
