+++ src/gas/shared/Response.ts
